Fix: Incorrect handling of asset libraries from Preferences
Asset libraries from the Preferences were not recognized as such. This
could cause various issues like the asset library not being refreshed
correctly on relevant changes. There was also a failed assert when
right-clicking the context menu of a custom brush asset, and the "Edit
Metadata" brush asset operator would be unavailable.

A pointer was unintentionally cast to a boolean, and instead null passed
for the pointer (because of a default argument value).