Nodes: improve "frame" search
This reduces the search weight of the frame node because it is rarely added
through the add menu. It's much easier to add it with the F key.

By reducing its search weight, the Frame output of the Scene Time node shows up
as first result, which is more desirable.

Pull Request: https://projects.blender.org/blender/blender/pulls/139709